cost-data-import-criteo/requirements.txt
# Function dependencies, for example:
# package>=version
requests
pandas
datetime
google-cloud-storage
google-api-python-client
oauth2client
criteo_marketing
